Factored out fee rate assertions and added to integration tests
