+++ dlc-test/src/test/scala/org/bitcoins/dlc/DLCFeeTestUtil.scala
@@ -1,0 +1,106 @@
+package org.bitcoins.dlc
+
+import org.bitcoins.core.currency.CurrencyUnit
+import org.bitcoins.core.protocol.transaction.Transaction
+import org.bitcoins.dlc.builder.DLCTxBuilder
+import org.scalatest.{Assertion, Assertions}
+
+object DLCFeeTestUtil extends Assertions {
+
+  def validateFees(
+      builder: DLCTxBuilder,
+      fundingTx: Transaction,
+      closingTx: Transaction): Assertion = {
+    val feeRate = builder.feeRate
+    val finalizer = builder.fundingTxBuilder.fundingTxFinalizer
+    val expectedFundingFee =
+      finalizer.offerFundingFee + finalizer.acceptFundingFee
+    val expectedClosingFee =
+      finalizer.offerFutureFee + finalizer.acceptFutureFee
+
+    val fundingInput =
+      (builder.offerFundingInputs ++ builder.acceptFundingInputs)
+        .map(_.output.value)
+        .sum
+    val fundingOutput = fundingTx.outputs.map(_.value).sum
+    val actualFundingFee = fundingInput - fundingOutput
+
+    val closingInput = fundingTx.outputs.head.value
+    val closingOutput = closingTx.outputs.map(_.value).sum
+    val actualClosingFee = closingInput - closingOutput
+
+    /** Actual Fee Rate = Actual Fee / Ceil(Actual Weight / 4.0)
+      * Expected Fee Rate = Expected Fee / Ceil(Expected Weight / 4.0)
+      *
+      * Expected Fee = Actual Fee (yay!)
+      *
+      * Expected Weight - #sigs - 4 <= Actual Weight <= Expected Weight
+      * The right comparison is true because Expected Weight is designed to be an upper bound.
+      * The left comparison is true because the possible savings are from one weight being saved per
+      * signature, and 1 vbyte = 4 weight being saved if rounding works out well between both parites.
+      *
+      * Because of these comparisons, we can derive
+      *
+      * Lower Bound Fee Rate = Actual Fee / (Ceil((Actual Weight + #sigs)/4.0) + 1)
+      * Upper Bound Fee Rate = Actual Fee / Ceil((Actual Weight + #sigs)/4.0)
+      *
+      * So that these two fee rates correspond to vbyte amounts 1 apart and allowing for signature
+      * size variation. This function asserts:
+      * Lower Bound Fee Rate <= Actual Fee Rate <= Upper Bound Fee Rate
+      */
+    def feeRateBounds(
+        tx: Transaction,
+        actualFee: CurrencyUnit,
+        missingOutputBytes: Long = 0,
+        numSignatures: Int = 2): (Double, Double) = {
+      val vbytesLower =
+        Math.ceil((tx.weight + numSignatures) / 4.0) + missingOutputBytes
+      val vbytesUpper =
+        vbytesLower + 1
+
+      // Upper VBytes => Lower fee rate
+      val lowerBound = actualFee.satoshis.toLong / vbytesUpper
+      // Lower VBytes => Upper fee rate
+      val upperBound = actualFee.satoshis.toLong / vbytesLower
+      (lowerBound, upperBound)
+    }
+
+    val (actualFundingFeeRateLower, actualFundingFeeRateUpper) =
+      feeRateBounds(fundingTx, actualFundingFee)
+
+    val (actualClosingFeeRateLower, actualClosingFeeRateUpper) =
+      feeRateBounds(closingTx, actualClosingFee)
+
+    val offerOutputBytes =
+      9 + builder.offerFinalAddress.scriptPubKey.asmBytes.length
+    val acceptOutputBytes =
+      9 + builder.acceptFinalAddress.scriptPubKey.asmBytes.length
+
+    val (actualClosingFeeRateWithoutOfferLower,
+         actualClosingFeeRateWithoutOfferUpper) =
+      feeRateBounds(closingTx, actualClosingFee, offerOutputBytes)
+
+    val (actualClosingFeeRateWithoutAcceptLower,
+         actualClosingFeeRateWithoutAcceptUpper) =
+      feeRateBounds(closingTx, actualClosingFee, acceptOutputBytes)
+
+    def feeRateBetweenBounds(
+        lowerBound: Double,
+        upperBound: Double): Boolean = {
+      feeRate.toLong >= lowerBound && feeRate.toLong <= upperBound
+    }
+
+    assert(actualFundingFee == expectedFundingFee)
+    assert(actualClosingFee == expectedClosingFee)
+    assert(
+      feeRateBetweenBounds(actualFundingFeeRateLower,
+                           actualFundingFeeRateUpper))
+    assert(
+      feeRateBetweenBounds(actualClosingFeeRateLower,
+                           actualClosingFeeRateUpper) ||
+        feeRateBetweenBounds(actualClosingFeeRateWithoutOfferLower,
+                             actualClosingFeeRateWithoutOfferUpper) ||
+        feeRateBetweenBounds(actualClosingFeeRateWithoutAcceptLower,
+                             actualClosingFeeRateWithoutAcceptUpper))
+  }
+}
